addressing review comments, while waiting for the config changes to be merged with master
